rename helper.h --> inline_helper.h
